할인 요금 구하기

+++ src/main/java/io/studyobject/movie/Money.java
@@ -8,7 +8,7 @@ public class Money {
     private final BigDecimal amount;
 
     public static Money wons(BigDecimal amount) {
-        return new Money(BigDecimal.valueOf(amount));
+        return new Money(amount);
     }
 
     public static Money wons(double amount) {
